* refactor (step3) : CarRaceController 제거
- CarRaceController기능을 CarRacing으로 통합
- CarManagement 불필요한 메서드 삭제, getCarsList 네임 변경
